logout move to palette
