+++ src/components/Questions/Palette/QuestionPalette.vue
@@ -79,6 +79,18 @@
           </div>
       </div>
     </div>
+    <div
+      v-if="showPortalLogout"
+      class="mt-8 border-t border-gray-200 pt-4"
+      data-test="palette-logout"
+    >
+      <icon-button
+        :titleConfig="portalLogoutTitleConfig"
+        :buttonClass="portalLogoutButtonClass"
+        @click="$emit('logout')"
+        data-test="portalLogoutButton"
+      />
+    </div>
   </div>
 </template>
 
@@ -91,6 +103,7 @@ import Review from "./Review.vue";
 import PaletteItem from "./Item.vue";
 import InstructionPage from "@/components/InstructionPage.vue";
 import { TimeLimit, questionSetPalette, quizTitleType, testFormat } from "@/types";
+import IconButton from "@/components/UI/Buttons/IconButton.vue";
 import { defineComponent, computed, PropType, reactive } from "vue";
 
 export default defineComponent({
@@ -101,7 +114,8 @@ export default defineComponent({
     Neutral,
     Review,
     PaletteItem,
-    InstructionPage
+    InstructionPage,
+    IconButton
   },
   props: {
     hasQuizEnded: {
@@ -144,6 +158,14 @@ export default defineComponent({
     isOmrMode: {
       type: Boolean,
       default: false,
+    },
+    showPortalLogout: {
+      type: Boolean,
+      default: false,
+    },
+    portalLogoutLabel: {
+      type: String,
+      default: "Logout",
     },
   },
   setup(props, context) {
@@ -193,6 +215,15 @@ export default defineComponent({
     const showInstructionButton = computed(() => state.showInstructions == true)
     const showPaletteButton = computed(() => state.showPalette == true)
 
+    const portalLogoutTitleConfig = computed(() => ({
+      value: props.portalLogoutLabel || "Logout",
+      class: "text-red-700 text-xs bp-500:text-sm font-semibold",
+    }));
+
+    const portalLogoutButtonClass = computed(() =>
+      "bg-red-50 hover:bg-red-100 ring-red-100 p-2 px-3 bp-500:px-4 rounded-lg shadow-sm border border-red-200 w-full"
+    );
+
     const toggleInstructionsButtonClass = computed(() => [
       {
         "bg-primary": props.isOmrMode || !showInstructionButton.value,
@@ -222,8 +253,10 @@ export default defineComponent({
       togglePaletteButtonClass,
       legendErrorText,
       legendNeutralText,
+      portalLogoutTitleConfig,
+      portalLogoutButtonClass
     };
   },
-  emits: ["navigate"],
+  emits: ["navigate", "logout"],
 });
 </script>
