+++ src/components/Questions/QuestionModal.vue
@@ -13,12 +13,9 @@
       :userId="userId"
       :quizType="quizType"
       :displayId="displayId"
-      :showPortalLogout="showPortalLogout"
-      :portalLogoutLabel="portalLogoutLabel"
       @time-limit-warning="displayTimeLimitWarning"
       @end-test="endTest"
       @end-test-by-time="endTestByTime"
-      @logout="$emit('logout')"
       data-test="header"
     ></Header>
     <InfoBar
@@ -26,9 +23,6 @@
       :title="title"
       :userId="userId"
       :displayId="displayId"
-      :showPortalLogout="showPortalLogout"
-      :portalLogoutLabel="portalLogoutLabel"
-      @logout="$emit('logout')"
       data-test="infoBar"
     />
     <div
@@ -113,8 +107,11 @@
       :maxMarks="maxMarks"
       :numQuestions="numQuestions"
       :quizTimeLimit="quizTimeLimit"
+      :showPortalLogout="showPortalLogout"
+      :portalLogoutLabel="portalLogoutLabel"
       class="fixed left-0 top-20 h-[calc(100vh-5rem)] w-full sm:w-2/3 lg:w-1/2 xl:w-1/3 z-50 bg-white overflow-y-auto"
       @navigate="navigateToQuestion"
+      @logout="$emit('logout')"
       data-test="questionPalette"
     >
     </QuestionPalette>
@@ -712,15 +709,6 @@ For final submission, click the End Test button again.`,
       }
     }
 
-    const portalLogoutTitleConfig = computed(() => ({
-      value: props.portalLogoutLabel || "Logout",
-      class: "text-white text-sm bp-500:text-md lg:text-lg xl:text-xl font-bold",
-    }));
-
-    const portalLogoutButtonClass = computed(() =>
-      "bg-gray-500 ring-gray-500 p-2 px-4 bp-500:p-4 bp-500:px-6 rounded-lg sm:rounded-2xl shadow-xl hover:bg-gray-600 focus:outline-none focus-visible:ring-2 focus-visible:ring-offset-1 focus-visible:ring-gray-300"
-    );
-
     return {
       ...toRefs(state),
       questionOptionSelected,
@@ -751,9 +739,7 @@ For final submission, click the End Test button again.`,
       numericalAnswerUpdated,
       isQuestionTypeMatrixSubjective,
       timeLimitWarningThreshold,
-      displayTimeLimitWarning,
-      portalLogoutTitleConfig,
-      portalLogoutButtonClass
+      displayTimeLimitWarning
     }
   },
   emits: [
